many of the slots with the from_to index are not used because there are no possible move between the squares. So i calculate now at startup an mapping array to reduce the size of cmh array. This increase the size of this table now only by 75% (1.5 MB per thread). The main history is also reduced (by 4 KB per thread) because it used the same index. But the mapping array adds 8 KB globaly to memory usage. Bench: 3606853

+++ src/movepick.cpp
@@ -22,7 +22,7 @@
 
 #include "movepick.h"
 
-int fromToIndex[SQUARE_NB][SQUARE_NB];
+int16_t fromToIndex[SQUARE_NB][SQUARE_NB];
 
 int from_to(Move m) {
   return fromToIndex[from_sq(m)][to_sq(m)];
